Add deep clone to metaSteps (#12)
As codeceptjs works on the metaStep object in the background it happens
temporarily that the metaStep looses the success or failed status.
It has the status "queued" then. This leads to errors on the
reportportal server, which cannot handle the status "queued".

With the deepClone the metaStep in the reporter is uncoupled
from codeceptjs.